:test_tube: test (vn_entropy/mutua_info): Move tests into the measurements/ folder. (#3544)
